% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso.R
\name{robsel.glasso}
\alias{robsel.glasso}
\title{Fit Graphical Lasso with RobSel}
\usage{
robsel.glasso(x, alpha = 0.9, B = 200, ...)
}
\arguments{
\item{x}{A \code{n}-by-\code{p} data matrix}

\item{alpha}{Prespecified confidence level. Default 0.9}

\item{B}{Number of bootstrap sample. Default 200}

\item{...}{Optional arguments passed on to glasso.}
}
\value{
A list with components:
\item{alpha}{A list of prespecified confidence level}
\item{lambda}{A list of estimate regularization parameter for Graphical Lasso}
\item{Omega}{A list of estimated inverse covariance matrix}
\item{Sigma}{A list of estimated covariance matrix}
}
\description{
Fit  Graphical Lasso with estimate regularization parameter from Robust Selection
}
\note{
Each item in each compenent corresponds to a prespecified level alpha.
}
\examples{
set.seed(17)
library(robsel)
x <-matrix(rnorm(50*20),ncol=20)

#Use Graphical Lasso with estimate regularization parameter lambda from RobSel
fit <- robsel.glasso(x = x, alpha = 0.9, B = 200)

}
\references{
P Cisneros-Velarde, A Petersen and S-Y Oh (2020). Distributionally Robust Formulation and Model Selection for the Graphical Lasso. Proceedings of the Twenty Third International Conference on Artificial Intelligence and Statistics.

Friedman, Jerome, Trevor Hastie, and Robert Tibshirani. 'Sparse inverse covariance estimation with the graphical lasso.' \emph{Biostatistics} 9.3 (2008): 432-441.

Meinshausen, Nicolai and Buhlmann, Peter. 2006. 'High-Dimensional Graphs and Variable Selection with the Lasso.' \emph{The Annals of Statistics}. JSTOR: 1436-1462.

Witten, Daniela M, Friedman, Jerome H, and Simon, Noah. 2011. 'New Insights and Faster computations for the Graphical Lasso.' \emph{Journal of Computation and Graphical Statistics}. Taylor and Francis: 892-900.
}
\seealso{
\link[robsel]{robsel} for Robust Selection algorithm, \link[glasso]{glasso} for Graphical Lasso algorithm.
}
